The MemArena code has way too many ifdefs. Split it up by platform.
