+++ Common/Common.vcxproj
@@ -292,7 +292,9 @@
     </ClCompile>
     <ClCompile Include="KeyMap.cpp" />
     <ClCompile Include="LogManager.cpp" />
-    <ClCompile Include="MemArena.cpp" />
+    <ClCompile Include="MemArenaAndroid.cpp" />
+    <ClCompile Include="MemArenaPosix.cpp" />
+    <ClCompile Include="MemArenaWin32.cpp" />
     <ClCompile Include="MemoryUtil.cpp" />
     <ClCompile Include="MipsEmitter.cpp" />
     <ClCompile Include="Misc.cpp" />
@@ -314,9 +316,6 @@
     <ClCompile Include="Vulkan\VulkanMemory.cpp" />
     <ClCompile Include="x64Analyzer.cpp" />
     <ClCompile Include="x64Emitter.cpp" />
-  </ItemGroup>
-  <ItemGroup>
-    <None Include="CMakeLists.txt" />
   </ItemGroup>
   <ItemGroup>
     <ProjectReference Include="..\ext\native\native.vcxproj">
