+++ Common/Common.vcxproj.filters
@@ -81,7 +81,6 @@
     <ClCompile Include="CPUDetect.cpp" />
     <ClCompile Include="FileUtil.cpp" />
     <ClCompile Include="LogManager.cpp" />
-    <ClCompile Include="MemArena.cpp" />
     <ClCompile Include="MemoryUtil.cpp" />
     <ClCompile Include="Misc.cpp" />
     <ClCompile Include="MsgHandler.cpp" />
@@ -132,9 +131,9 @@
     <ClCompile Include="Vulkan\VulkanMemory.cpp">
       <Filter>Vulkan</Filter>
     </ClCompile>
-  </ItemGroup>
-  <ItemGroup>
-    <None Include="CMakeLists.txt" />
+    <ClCompile Include="MemArenaPosix.cpp" />
+    <ClCompile Include="MemArenaWin32.cpp" />
+    <ClCompile Include="MemArenaAndroid.cpp" />
   </ItemGroup>
   <ItemGroup>
     <Filter Include="Crypto">
